grid-linking: separated box to a seperate file
